Add files via upload
This Java program showcases the usage of variable arguments (varargs) in method parameters. It defines a method sum that can take any number of integer arguments and calculates their sum. This allows for flexibility in method invocation, enabling the user to provide any number of integers to be summed up.